feat: add always-on filters option to AdminOptions 🎨
This commit introduces a new feature in the `AdminOptions` class by adding a `ra_filters_always_on` attribute. This allows developers to specify which filters should always be active on the admin interface, enhancing user experience by pre-filtering data as needed. The changes are implemented in the `handler.py` file, ensuring that filters specified in `ra_filters_always_on` are marked with the "alwaysOn" property.